all: move cppad{,cg} and casadi to the right directories
