+++ unittest/casadi-basic.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2019 INRIA
 //
 
-#include <pinocchio/math/casadi.hpp>
+#include <pinocchio/autodiff/casadi.hpp>
 #include <Eigen/Dense>
 
 #include <boost/variant.hpp> // to avoid C99 warnings
