+++ unittest/cppad-algo-derivatives.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2018-2019 CNRS INRIA
 //
 
-#include "pinocchio/math/cppad.hpp"
+#include "pinocchio/autodiff/cppad.hpp"
 
 #include "pinocchio/multibody/model.hpp"
 #include "pinocchio/multibody/data.hpp"
