+++ unittest/cppad-basic.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) 2018-2019 CNRS INRIA
 //
 
-#include "pinocchio/math/cppad.hpp"
+#include "pinocchio/autodiff/cppad.hpp"
 #include <cppad/speed/det_by_minor.hpp>
 
 #include <boost/variant.hpp> // to avoid C99 warnings
